+++ tests/Silverback.Integration.Tests.E2E/TestTypes/Messages/TestCommandOne.cs
@@ -1,0 +1,12 @@
+// Copyright (c) 2020 Sergio Aquilini
+// This code is licensed under MIT license (see LICENSE file for details)
+
+using Silverback.Messaging.Messages;
+
+namespace Silverback.Tests.Integration.E2E.TestTypes.Messages
+{
+    public class TestCommandOne : IIntegrationCommand
+    {
+        public string? Content { get; set; }
+    }
+}