+++ tests/Silverback.Integration.Tests/TestTypes/InboundLoggerSubstitute`1.cs
@@ -53,6 +53,10 @@ namespace Silverback.Tests.Integration.TestTypes
         {
         }
 
+        public void LogNullMessageSkipped(IRawInboundEnvelope envelope)
+        {
+        }
+
         public void LogAlreadyProcessed(IRawInboundEnvelope envelope)
         {
         }